requirements.txt
PyYAML==6.0
requests==2.22.0
tweepy==3.10.0